+++ stepik-union/src/test/java/org/stepik/plugin/utils/DirectivesUtilsTest.java
@@ -1,11 +1,10 @@
-package org.stepik.plugin.actions;
+package org.stepik.plugin.utils;
 
 import org.jetbrains.annotations.NotNull;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.stepik.plugin.collective.SupportedLanguages;
-import org.stepik.plugin.utils.DirectivesUtils;
 
 import java.io.InputStream;
 import java.util.HashMap;
